Angular-s4

+++ src/app/heroes/heroes.component/heroes.component.ts
@@ -35,8 +35,10 @@ export class HeroesComponent implements OnInit {
   }
 
   delete(hero: Hero): void {
-    this.heroes = this.heroes.filter(h => h !== hero);
-    this.heroService.deleteHero(hero).subscribe();
+    if(hero.id > 3){
+      this.heroes = this.heroes.filter(h => h !== hero);
+      this.heroService.deleteHero(hero).subscribe();
+    }
   }
 
 }
